split tests into generators and generators_base; add explicit compat for new params
